DDR4 refactor wip
