+++ src/DRAMPower/CMakeLists.txt
@@ -16,6 +16,8 @@ add_library(DRAMPower
     DRAMPower/memspec/MemSpecLPDDR5.cpp
     DRAMPower/standards/ddr4/core_calculation_DDR4.cpp
     DRAMPower/standards/ddr4/DDR4.cpp
+    DRAMPower/standards/ddr4/DDR4Core.cpp
+    DRAMPower/standards/ddr4/DDR4Interface.cpp
     DRAMPower/standards/ddr4/interface_calculation_DDR4.cpp
     DRAMPower/standards/ddr5/core_calculation_DDR5.cpp
     DRAMPower/standards/ddr5/DDR5.cpp
@@ -30,6 +32,7 @@ add_library(DRAMPower
     DRAMPower/util/extensions.cpp
     DRAMPower/util/pin.cpp
     DRAMPower/util/dbi.cpp
+    DRAMPower/util/ImplicitCommandHandler.cpp
 )
 
 target_include_directories(DRAMPower PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
@@ -42,5 +45,10 @@ target_link_libraries(DRAMPower
 PUBLIC 
     DRAMUtils::DRAMUtils
 )
+
+# Add test functions
+if(DRAMPOWER_BUILD_TESTS)
+    target_compile_definitions(DRAMPower PUBLIC DRAMPOWER_TESTING)
+endif()
 
 add_library(DRAMPower::DRAMPower ALIAS DRAMPower)
