+++ src/DRAMPower/DRAMPower/data/stats.h
@@ -32,7 +32,7 @@ namespace DRAMPower
 			uint64_t powerDownPre = 0;
 			uint64_t selfRefresh = 0;
 			uint64_t deepSleepMode = 0;
-			uint64_t activeTime() { return act; };
+			uint64_t activeTime() const { return act; };
 			//uint64_t prechargeTime;
 		} cycles;
